ShoghlanahProject/requirements.txt
Django==1.4
PIL==1.1.7
South==0.7.6
ajaxuploader==0.2.4
argparse==1.2.1
boto==2.6.0
distribute==0.6.28
dj-database-url==0.2.0
-e git+https://github.com/toastdriven/django-haystack.git@58f4eddbf1044a2a5663de470301c0af7a957b13#egg=django_haystack-dev
django-openid-auth==0.4
django-session-security==2.0.3
django-storages==1.1.5
gunicorn==0.15.0
httplib2==0.7.5
oauth2==1.5.211
psycopg2==2.4.5
pyelasticsearch==0.2
python-openid==2.2.5
requests==0.14.0
sorl-thumbnail==11.12
stripe==1.7.5
tweepy==2.0
wsgiref==0.1.2
